api-server/requirements.txt
django>=5.0,<6.0
django-cors-headers>=4.3.0,<5.0
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0,<1.0
pymongo>=4.8.0,<5.0
python-dotenv>=1.0.0,<2.0
PyJWT>=2.9.0,<3.0
bcrypt>=3.2.0,<4.0.0
qrcode[pil]>=7.4.0,<8.0
python-multipart>=0.0.9,<1.0
websockets>=12.0,<13.0
